chore: ci cleanup
move to gh actions
add npm linting
fix tests for new tap
