+++ .eslintrc.js
@@ -1,0 +1,12 @@
+const { readdirSync: readdir } = require('fs')
+
+const localConfigs = readdir(__dirname)
+  .filter((file) => file.startsWith('.eslintrc.local.'))
+  .map((file) => `./${file}`)
+
+module.exports = {
+  extends: [
+    '@npmcli',
+    ...localConfigs,
+  ],
+}
