--- .travis.yml
@@ -1,5 +1,0 @@
-language: node_js
-sudo: false
-node_js:
-  - "12"
-  - "10"
