+++ lib/tracker-base.js
