+++ lib/tracker-stream.js
