+++ package.json
@@ -2,37 +2,52 @@
   "name": "are-we-there-yet",
   "version": "1.1.5",
   "description": "Keep track of the overall completion of many disparate processes",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
-    "test": "standard && tap test/*.js"
+    "test": "tap",
+    "npmclilint": "npmcli-lint",
+    "lint": "eslint '**/*.js'",
+    "lintfix": "npm run lint -- --fix",
+    "posttest": "npm run lint",
+    "postsnap": "npm run lintfix --",
+    "preversion": "npm test",
+    "postversion": "npm publish",
+    "prepublishOnly": "git push origin --follow-tags",
+    "snap": "tap"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/npm/are-we-there-yet.git"
   },
-  "author": "Rebecca Turner (http://re-becca.org)",
+  "author": "GitHub Inc.",
   "license": "ISC",
   "bugs": {
     "url": "https://github.com/npm/are-we-there-yet/issues"
   },
   "homepage": "https://github.com/npm/are-we-there-yet",
   "devDependencies": {
-    "standard": "^11.0.1",
-    "tap": "^12.0.1"
+    "@npmcli/eslint-config": "^1.0.0",
+    "@npmcli/template-oss": "^1.0.2",
+    "eslint": "^7.32.0",
+    "eslint-plugin-node": "^11.1.0",
+    "tap": "^15.0.9"
   },
   "dependencies": {
     "delegates": "^1.0.0",
-    "readable-stream": "^2.0.6"
+    "readable-stream": "^3.6.0"
   },
   "files": [
-    "index.js",
-    "tracker-base.js",
-    "tracker-group.js",
-    "tracker-stream.js",
-    "tracker.js",
-    "CHANGES.md"
+    "bin",
+    "lib"
   ],
   "engines": {
     "node": ">=10"
-  }
+  },
+  "tap": {
+    "branches": 68,
+    "statements": 92,
+    "functions": 86,
+    "lines": 92
+  },
+  "templateVersion": "1.0.2"
 }
